BREAKING: Rename the following in snapshot.properites au.com.origin.snapshots.serializers.DeterministicJacksonSnapshotSerializer to au.com.origin.snapshots.jackson.serializers.DeterministicJacksonSnapshotSerializer au.com.origin.snapshots.serializers.JacksonSnapshotSerializer to au.com.origin.snapshots.jackson.serializers.JacksonSnapshotSerializer
