+++ gradle.properties
@@ -1,2 +1,2 @@
 group=io.github.origin-energy
-version=3.3.9-SNAPSHOT
+version=3.1.0-SNAPSHOT
