+++ java-snapshot-testing-junit5/src/test/resources/snapshot.properties
@@ -1,5 +1,5 @@
 serializer=au.com.origin.snapshots.serializers.ToStringSnapshotSerializer
-serializer.json=au.com.origin.snapshots.serializers.DeterministicJacksonSnapshotSerializer
+serializer.json=au.com.origin.snapshots.jackson.serializers.DeterministicJacksonSnapshotSerializer
 comparator=au.com.origin.snapshots.comparators.PlainTextEqualsComparator
 reporters=au.com.origin.snapshots.reporters.PlainTextSnapshotReporter
 snapshot-dir=__snapshots__
